refactor: memos 폴더 이름 retrospect 로 변경 (#68)
