+++ src/app/groups/[id]/_components/MemoList.tsx
@@ -4,7 +4,7 @@ import Button from '@/components/Button'
 import { Icon } from '@/components/Icon'
 import Dropdown from '@/components/Dropdown'
 import AuthorInfo from '@/components/AuthorInfo'
-import IconWithButton from '@/app/memos/[id]/_components/IconWithButton'
+import IconWithButton from '@/app/retrospects/[id]/_components/IconWithButton'
 import { RetrospectList } from '../_types'
 import { URL_PATH } from '@/consts/urls'
 import { useRouter } from 'next/navigation'
