+++ src/app/retrospects/[id]/_components/GroupInfo.tsx
